Move backpressure to write pipeline
